+++ src/test/java/racingcar/controller/RacingGameControllerTest.java
@@ -14,7 +14,6 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import racingcar.dto.RequestDto;
 
 @DisplayName("post 요청 시 상태코드 200 반환 및 반환된 json 형식 데이터 확인")
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
